renamed api-server to ros2-bridge; remove trajectory server plugin
