+++ packages/dashboard/e2e/scripts/test-e2e.js
@@ -33,7 +33,7 @@ const wdioArgs = process.argv
 
 concurrently(
   [
-    'cd .. && npm run start:api',
+    'cd .. && npm run ros2-bridge',
     'cd .. && npm run start:auth',
     'npm:start:react',
     `node auth-ready.js && wdio ${wdioArgs}`,
