+++ packages/ros2-bridge/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "api-server",
+  "name": "ros2-bridge",
   "version": "0.0.1",
-  "description": "web + data server for romi-dashboard",
+  "description": "bridges ros2 messages to websocket/msgpack",
   "main": "src/index.js",
   "scripts": {
     "preinstall": "node ../../scripts/install.js",
@@ -17,7 +17,7 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/osrf/romi-dashboard",
-    "directory": "packages/api-server"
+    "directory": "packages/ros2-bridge"
   },
   "dependencies": {
     "@msgpack/msgpack": "^2.0.0",
@@ -40,7 +40,7 @@
     "typescript": "^4.0.2"
   },
   "bin": {
-    "api-server": "src/index.js"
+    "ros2-bridge": "src/index.js"
   },
   "files": [
     "src/**/*.js",
