+++ packages/ros2-bridge/src/authenticator.ts
