+++ packages/ros2-bridge/src/logger.ts
@@ -1,8 +1,10 @@
 import * as winston from 'winston';
 
-export interface CustomLogger extends winston.Logger {
+interface CustomLogger extends winston.Logger {
   child: (metadata: { tag: unknown }) => CustomLogger;
 }
+
+export { CustomLogger as Logger };
 
 export const logger: CustomLogger = winston.createLogger({
   format: winston.format.combine(
